move time entry validations to contract
